+++ analysis/reanalyze/src/ExnLib.ml
@@ -21,7 +21,14 @@ let raisesLibTable : (Name.t, Exceptions.t) Hashtbl.t =
   in
   let beltMap = [("getExn", [notFound]); ("getOrThrow", [notFound])] in
   let beltMutableMap = beltMap in
-  let beltMutableQueue = [("peekExn", [notFound]); ("popExn", [notFound])] in
+  let beltMutableQueue =
+    [
+      ("peekExn", [notFound]);
+      ("peekOrThrow", [notFound]);
+      ("popExn", [notFound]);
+      ("popOrThrow", [notFound]);
+    ]
+  in
   let beltSet = [("getExn", [notFound]); ("getOrThrow", [notFound])] in
   let beltMutableSet = beltSet in
   let beltOption = [("getExn", [notFound])] in